Fix #104 "Review page getting stuck loading" (thanks ellie)
It looks like I was halfway through refactoring a view template and an
element with class="msg" disappeared. Another class was added to the
root element for the connection error message class="uiFcErrorMsg"..
but a grep reveals it is not used anywhere.

Then a look into AjaxQueue setErrorDialog() code as well as its
documentation confirms that it wants a child element of class "msg".
We fix that and also changes the class to a more unique class name
"uiFcAjaxError_msg" which makes it easier to grep the use of this class
("msg" is far too generic).

Pretty simple bug caused by refactoring a while ago.