Enable the use of relative paths in Juicer config file.
A stepping stone towards simplifying setup for contributors,
this allows using ./rel/ative paths in the Juicer config file.

Specifically, any Juicer %CONSTANT% used in =require from
directives can use a relative path, and will be noramlized 
to an absolute paht with SF_ROOT_DIR as the root of the
project.