Improve design and fix paypal.me with correct PayPal donate button code
As part of the responsive update early 2017, I switched the PayPal donate
button to a paypal.me page. This is incorrect however as paypal.me is meant
for personal payments (friends / family), and PayPal fee wasnt processed.

The PayPal donate button has been reverted to the original code. In addition
Ì noticed PayPal checkout adds a "recurring" option checkbox, so this
simplifies the design.