Add rtkValidators::sanitizeCJKUnifiedUCS() very handy for ajax requests
sanitizeXXX methods do both validation AND type casting if necessary to the expected type (typically, numbers that should be int). Simplifies the validation code in ajax request handlers where we don't expect the user to mess with the content of the POST body. If it's not valid, throw an exception.