(build) phase out Core.log() calls from all the js/vue under lib/front/vue/
/lib/front/vue/    is where the Vue build with Webpack 3 resides

Core.log() is a legacy method that used to disable logs in production,
but it was still output in the minified js.

Since updating uglifyjs we could use "pure_funcs" to drop Core.log()
calls but that seems very dangerous since a npm package could match
that function.


TLDR

Use console.log() from here on, with "drop_console" option of uglifyjs
which is safer. We can also use .warn() and .error() and they will be
stripped from production.