Juicer now parses Sass-like variables, add theme.config.css
Juicer used to support variables (interpolated in BOTH css and js, in
fact Juicer doesn't care what the file types are, for the most part).

The syntax was %VAR_NAME%. The syntax was updated to one
documented in postcss-advanced-variables plugin, just in case we
ditch the legacy build some day (if ever).

But also, we move the css variables to a css include, which may enable
creating a light / dark theme some day.