Add a `entry-common` so we don't include study-specific css/js on all pages
We were using `entry-study` as a "base bundle" which inits the desktop &
mobile navs, but also includes css/js specific to Study pages on all
other pages.

Instead add a `entry-common`, which inits the mobile/desktop nav across
the site, but also acts as a "common" chunk which groups all code
Rollup identifies as shared between entries.

For now to keep things simple, just include entry-study on the Profile
page as it's the only "non study" module/action which uses My Stories
css.