Fix #207 : Unable to close the Edit Story dialog...
This doesn't actually fix why clicking the masks no longer dismisses
the Edit Story or Dict dialogs after a while.

The bug may be in my code (likely), or related to YUI2 dialogs. Since
those dialogs are obsolete and should be replaced with Vue 3 components
in coming months... it might not be worth fixing.

This "fix" then adds a large "Close" button to the Edit Story dialog,
even on desktop.

This lets the user always dismiss the Edit Story dialog, even if the
mask area stops responding to clicks.