(apache): fix mod_rewrite rule to match Vite build's hashed filenames
Sets far future expire headers to the Vite build assets.

Other css/js can be versioned by using similar pattern:

    /path/to/file.12345678.(css|js)

NOTES

The old cache.php used to remove the hash, so that filenames on the
server don't need to change. That doesn't work with Vite's default
output, but then again we rarely need to change the filenames anyway
-- so instead we'll update the filenames for the non-Vite build files
to match the rewrite rule.