(ReviewsPeer): add foolproof check for invalid leitnerbox
This check also prevents an "out of memory" error due to unexpected
loop logic following.

This will still prevent accessing the SRS chart; which is probably
better if a flashcard upate somehow messed the leitnerbox value -- to
prevent further reviews from happening.

But this will output errors in error_log().

Part of #232 resolution