Always clear kanji from learned list when a card is reviewed
No idea why this was ever conditional to the API (KanjiRyokucha)...

The list is no longer cleared at end of SRS review. Learned kanji
remain Indefinitely unless user explicitly clears the list, or during
review.

If kanji is forgotten, it needs to be re-learned, so clear flag.

If kanji is remembered, it is no longer in the Restudy list, clear flag.